Made a bunch of file sizes be 64-bit values, made us avoid using 32-bit stat functions, added filestat.hpp/cc and get_file_size function.

+++ src/progress/progress.hpp
@@ -73,7 +73,7 @@ private:
     void on_ring();
 
     FILE *file;
-    int file_size;
+    int64_t file_size;
 
     progress_bar_t progress_bar;
